+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Please also have a look at our
 
 ### Changed
 
+- Cleanup extra whitespace in css selector (#1398)
 - The array keys passed to `DeclarationBlock::setSelectors()` are no longer
   preserved (#1407)