+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Please also have a look at our
 
 ### Added
 
+- Provide line number in exception message for mismatched parentheses in
+  selector (#1435)
 - Add support for CSS container queries (#1400)
 
 ### Changed